fix bug

+++ Strategy/sell_stock_by_pct.py
@@ -17,7 +17,7 @@ def sell_by_pct(stock_list, stock_inventory, pct_threshold_to_sell=0.02):
 
         stock_current_price = float(stock_inventory[stock_symbol]['price'])
         stock_average_buy_price = float(stock_inventory[stock_symbol]['average_buy_price'])
-        stock_holding_units = float(stock_inventory[stock_symbol]['quantity'])
+        stock_holding_units = int(stock_inventory[stock_symbol]['quantity'])
 
         if stock_current_price / stock_average_buy_price - 1 > pct_threshold_to_sell:
